fixes #269
